refactor(test): split scroll-edge-state suite further, tighten unreachable branch
Splits container-selection.test.ts (562 LOC, over the 500-LOC extract
tripwire) into four narrower suites by durable question: container
existence, hidden-edge decisions, target resolution, and broad selection.
Adds a scrollNode() fixture builder to cut inline node-literal repetition
across the suite.

Removes the "captureNodes resolves to undefined" test: the real contract
(both production call sites, and captureScrollEdgeState's own captureNodes
type) always resolves to a real array, so this exercised an unreachable
branch behind a type-unsound cast. Tightens analyzeScrollEdgeState's
parameter type to drop the now-provably-unreachable `| undefined`.

Simplifies buildScrollContainerScope to filter out non-string
identifier/label fields instead of mapping them to an empty-string
placeholder — behaviorally identical (the placeholder always failed
isUsefulScope's length check anyway) but removes a StringLiteral mutation
surface that a synthetic fallback could never legitimately kill without
coupling a test to Stryker's specific replacement text.

Addresses further review feedback on PR #1474 (#1455).
